fix: Remove the inspector module
In the first version of Boltx, all messages were created in separate
modules per version, and this module was used to find messages by
searching in the modules. The main idea was to avoid having a static
list of versions and instead find them dynamically.

However, this idea was discarded due to the complexity and the very
frequent changes in messages per Bolt version. Nevertheless, the file
implementing this idea was not removed.